Rework minimal example
I am trying to use as few abstraction layers as possible and just turn
on the LED on the Pico board. No HALs, no BSPs. This is the best I could
come up with so far. I could probably replace the delay function and
remove the direct dependency to cortex-m.