Temporarily disable the pwm example
The reset performed by `cargo flash` does not reset the clocks but
instead leaves the clock configuration as the pwm example sets it up.
This means that the gpio examples do not have the original behavior
(using the default ROSC clock) when running them.

A Power-on reset seems to fix the issue so I have to figure out a way
to do that. Currently looking into getting a USB Hub where I can
control the USB port power using `uhubctl`, and use one of those ports
for powering the Raspberry Pi Pico. In the future I will most likely
get a programmable power supply.