Rework pipeline and add delay to power supply script
Flashing the RPI and not doing a power cycle afterwards risks using the
configuration of whatever was previously flashed.

To be more precise, after flashing the gpio_toggle_unsafe example, some
of the clock configuration from the pwm_with_hal example still
persisted. This was visible in the Saleae capture, the number of samples
was greater than usual.

Doing a power cycle returns all registers (including those associated
with the clock configuration) to their default values, which is what we
want.